+++ routes.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"errors"
+
 	"github.com/xmidt-org/themis/key"
 	"github.com/xmidt-org/themis/token"
 	"github.com/xmidt-org/themis/xhealth"
@@ -73,7 +75,7 @@ type KeyRoutesIn struct {
 }
 
 func BuildKeyRoutes(in KeyRoutesIn) {
-	if in.Router != nil {
+	if in.Router != nil && in.Handler != nil {
 		in.Router.Handle("/keys/{kid}", in.Handler).Methods("GET")
 	}
 }
@@ -85,7 +87,7 @@ type IssuerRoutesIn struct {
 }
 
 func BuildIssuerRoutes(in IssuerRoutesIn) {
-	if in.Router != nil {
+	if in.Router != nil && in.Handler != nil {
 		in.Router.Handle("/issue", in.Handler).Methods("GET")
 	}
 }
@@ -97,9 +99,42 @@ type ClaimsRoutesIn struct {
 }
 
 func BuildClaimsRoutes(in ClaimsRoutesIn) {
-	if in.Router != nil {
+	if in.Router != nil && in.Handler != nil {
 		in.Router.Handle("/claims", in.Handler).Methods("GET")
 	}
+}
+
+// CheckServerRequirements is an fx.Invoke function that does post-configuration verification
+// that we have required servers.  The valid server configurations are:
+//
+//    Both keys and issuer present.  Claims is optional in this case
+//    Neither keys or issuer present.  Claims is required in this case
+//
+// Any other arrangements results in an error.
+func CheckServerRequirements(k KeyRoutesIn, i IssuerRoutesIn, c ClaimsRoutesIn) error {
+	if k.Router != nil && i.Router != nil {
+		// all good ... no need to check anything else
+		return nil
+	}
+
+	if k.Router == nil && i.Router == nil {
+		if c.Router == nil {
+			return errors.New("A claims server is required if no keys or issuer server is configured")
+		}
+
+		// Only a claims server is allowed
+		return nil
+	}
+
+	if k.Router != nil {
+		return errors.New("If a keys server is configured, an issuer server must be configured")
+	}
+
+	if i.Router != nil {
+		return errors.New("If an issuer server is configured, a keys server must be configured")
+	}
+
+	return nil
 }
 
 type MetricsRoutesIn struct {
@@ -109,7 +144,9 @@ type MetricsRoutesIn struct {
 }
 
 func BuildMetricsRoutes(in MetricsRoutesIn) {
-	in.Router.Handle("/metrics", in.Handler).Methods("GET")
+	if in.Router != nil && in.Handler != nil {
+		in.Router.Handle("/metrics", in.Handler).Methods("GET")
+	}
 }
 
 type HealthRoutesIn struct {
@@ -119,5 +156,7 @@ type HealthRoutesIn struct {
 }
 
 func BuildHealthRoutes(in HealthRoutesIn) {
-	in.Router.Handle("/health", in.Handler).Methods("GET")
+	if in.Router != nil && in.Handler != nil {
+		in.Router.Handle("/health", in.Handler).Methods("GET")
+	}
 }