+++ release_notes.md
@@ -1,1 +1,3 @@
-
+- Allow metrics and health servers to be disabled
+- Allow only a claims server to be configured
+- Require an issuer server if a keys server is configured, and vice versa